Ex No 15 A. Calculation of sag with supports are at unequal levels using scilab
15a) The towers of height 30 m and 90 m respectively support a transmission line
conductor at water crossing. The horizontal distance between the towers is 500 m. If the
tension in the conductor is 1600 kg find the minimum clearance of the conductor and
water and clearance mid-way between the supports. Weight of conductor is 1-5 kg/m.
Bases of the towers can be considered to be at water level.