Ex No 15 B. Calculation of sag with supports are at unequal levels using scilab
15b) A to line has a span of 275 m between level supports. The conductor has an
effective diameter of 1.96 cm and diameter and weights o.865 kg/m. its ultimate strength
is 8060kg. if the conductor has the coating of radial thickness 1.27 cm and is subjected
to a wind pressure of 3.9 gm/m 2 of projected area calculate sag for a safety factor 2
weight of 1 c of ice is 0.91 gm.